server/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
flask-limiter==3.8.0
pymongo==4.8.0
python-dotenv==1.0.1
argon2-cffi==23.1.0
email-validator==2.2.0
boto3==1.35.0
itsdangerous==2.2.0
